+++ cx-portal/src/components/shared/basic/ReleaseProcess/AppMarketCard/index.tsx
@@ -500,8 +500,8 @@ export default function AppMarketCard() {
                 errors,
               }}
               name="provider"
-              maxLength={15}
-              minLength={1}
+              maxLength={30}
+              minLength={3}
               pattern={Patterns.appMarketCard.appProvider}
               label={t('content.apprelease.appMarketCard.appProvider') + ' *'}
               rules={{
@@ -510,7 +510,7 @@ export default function AppMarketCard() {
                 )} ${t('content.apprelease.appReleaseForm.isMandatory')}`,
                 minLength: `${t(
                   'content.apprelease.appReleaseForm.minimum'
-                )} 5 ${t(
+                )} 3 ${t(
                   'content.apprelease.appReleaseForm.charactersRequired'
                 )}`,
                 pattern: `${t(
@@ -518,7 +518,7 @@ export default function AppMarketCard() {
                 )} A-Za-z0-9.:_- @&`,
                 maxLength: `${t(
                   'content.apprelease.appReleaseForm.maximum'
-                )} 40 ${t(
+                )} 30 ${t(
                   'content.apprelease.appReleaseForm.charactersAllowed'
                 )}`,
               }}