+++ cx-portal/src/components/shared/basic/ReleaseProcess/components/ReleaseProcessWrapper.tsx
@@ -43,6 +43,7 @@ import OfferContractAndConsent from '../OfferContractAndConsent'
 import { serviceReleaseActiveStep } from 'features/serviceManagement/slice'
 import { ReleaseProcessTypes } from 'features/serviceManagement/apiSlice'
 import OfferValidateAndPublish from '../OfferValidateAndPublish'
+import { getAssetBase } from 'services/EnvironmentService'
 
 interface ReleaseProcessWrapperType {
   headerTitle: string
@@ -62,7 +63,7 @@ export default function ReleaseProcessWrapper({
   headerTitle,
   processType,
   stepsList,
-  imagePath = '../../submit-app-background.png',
+  imagePath = `${getAssetBase()}/images/content/submit-app-background.png`,
   headerDescription,
   headerDescriptionComplete,
   yourCatenaXTeam,