+++ cx-portal/src/components/shared/templates/AdminBoard/AdminBoardElements.tsx
@@ -20,17 +20,14 @@
 
 import { CardDecision, PageSnackbar } from 'cx-portal-shared-components'
 import { useDispatch, useSelector } from 'react-redux'
-import { useNavigate } from 'react-router-dom'
 import { useTheme, CircularProgress } from '@mui/material'
-import { useTranslation } from 'react-i18next'
 import { show } from 'features/control/overlay'
 import './AdminBoard.scss'
 import {
   AppContent,
   useApproveRequestMutation,
 } from 'features/adminBoard/adminBoardApiSlice'
-import NoItems from '../NoItems'
-import { OVERLAYS } from 'types/Constants'
+import { OVERLAYS, PAGES } from 'types/Constants'
 import {
   currentErrorType,
   currentSuccessType,
@@ -38,33 +35,72 @@ import {
   setSuccessType,
 } from 'features/adminBoard/slice'
 import { SuccessErrorType } from 'features/admin/appuserApiSlice'
+import NoItems from 'components/pages/NoItems'
+import {
+  ServiceContent,
+  useApproveServiceRequestMutation,
+} from 'features/adminBoard/serviceAdminBoardApiSlice'
+import { useState } from 'react'
 
-export default function AdminBoardElements({ apps }: { apps?: AppContent[] }) {
+export default function AdminBoardElements({
+  apps,
+  onClick,
+  type,
+  successApproveMsg,
+  errorApproveMsg,
+  successDeclineMsg,
+  errorDeclineMsg,
+}: {
+  apps?: AppContent[] | ServiceContent[]
+  onClick: (appId: string) => void
+  type?: string
+  successApproveMsg?: string
+  errorApproveMsg?: string
+  successDeclineMsg?: string
+  errorDeclineMsg?: string
+}) {
   const dispatch = useDispatch()
-  const navigate = useNavigate()
   const theme = useTheme()
-  const { t } = useTranslation()
-
   const [approveRequest] = useApproveRequestMutation()
+  const [approveServiceRequest] = useApproveServiceRequestMutation()
   const isDecisionSuccess = useSelector(currentSuccessType)
   const isDecisionError = useSelector(currentErrorType)
+  const [actionApprove, setActionApprove] = useState<boolean>(false)
 
   if (apps && apps.length === 0) {
     return <NoItems />
   }
 
   const handleApprove = async (appId: string) => {
-    await approveRequest(appId)
-      .unwrap()
-      .then(() => {
-        dispatch(setSuccessType(true))
-      })
-      .catch((error) => dispatch(setErrorType(true)))
+    setActionApprove(true)
+    if (type === PAGES.SERVICEADMINBOARD_DETAIL) {
+      await approveServiceRequest(appId)
+        .unwrap()
+        .then(() => {
+          dispatch(setSuccessType(true))
+        })
+        .catch((error) => dispatch(setErrorType(true)))
+    } else {
+      await approveRequest(appId)
+        .unwrap()
+        .then(() => {
+          dispatch(setSuccessType(true))
+        })
+        .catch((error) => dispatch(setErrorType(true)))
+    }
   }
 
   const onAlertClose = () => {
     dispatch(setSuccessType(false))
     dispatch(setErrorType(false))
+  }
+
+  const getDescription = () => {
+    if (actionApprove) {
+      return isDecisionSuccess ? successApproveMsg : errorApproveMsg
+    } else {
+      return isDecisionSuccess ? successDeclineMsg : errorDeclineMsg
+    }
   }
 
   return (
@@ -75,22 +111,23 @@ export default function AdminBoardElements({ apps }: { apps?: AppContent[] }) {
         severity={
           isDecisionSuccess ? SuccessErrorType.SUCCESS : SuccessErrorType.ERROR
         }
-        description={
-          isDecisionSuccess
-            ? t('content.adminBoard.successMsg')
-            : t('content.adminBoard.errorMsg')
-        }
+        description={getDescription()}
         showIcon={true}
         autoClose={true}
       />
       {apps && apps.length ? (
         <CardDecision
           items={apps}
-          onDelete={(appId: string) =>
-            dispatch(show(OVERLAYS.DECLINE_ADMINBOARD, appId))
-          }
+          onDelete={(appId: string) => {
+            setActionApprove(false)
+            if (type === PAGES.SERVICEADMINBOARD_DETAIL) {
+              dispatch(show(OVERLAYS.SERVICE_DECLINE_ADMINBOARD, appId))
+            } else {
+              dispatch(show(OVERLAYS.APP_DECLINE_ADMINBOARD, appId))
+            }
+          }}
           onApprove={(appId: string) => handleApprove(appId)}
-          onClick={(appId: string) => navigate(`/adminboarddetail/${appId}`)}
+          onClick={onClick}
         />
       ) : (
         <div className="loading-progress">